requirements.txt
Package            Version
------------------ -----------
asgiref            3.8.1
blinker            1.9.0
certifi            2025.4.26
charset-normalizer 3.4.2
click              8.2.1
cloudinary         1.44.0
colorama           0.4.6
contourpy          1.3.2
cycler             0.12.1
Django             5.2.1
filelock           3.18.0
Flask              3.1.1
fonttools          4.58.0
fsspec             2025.5.0
idna               3.10
itsdangerous       2.2.0
Jinja2             3.1.6
kiwisolver         1.4.8
MarkupSafe         3.0.2
matplotlib         3.10.3
mpmath             1.3.0
networkx           3.4.2
numpy              2.2.6
opencv-python      4.11.0.86
packaging          25.0
pandas             2.2.3
pillow             11.2.1
pip                24.0
psutil             7.0.0
py-cpuinfo         9.0.0
pyparsing          3.2.3
python-dateutil    2.9.0.post0
python-dotenv      1.1.0
pytz               2025.2
PyYAML             6.0.2
requests           2.32.3
scipy              1.15.3
setuptools         65.5.0
six                1.17.0
sqlparse           0.5.3
sympy              1.14.0
torch              2.7.0
torchvision        0.22.0
tqdm               4.67.1
typing_extensions  4.13.2
tzdata             2025.2
ultralytics        8.3.143
ultralytics-thop   2.0.14
urllib3            2.4.0
Werkzeug           3.1.3
